refactor: decouple client code and types
