+++ tee-worker/omni-executor/evm-aggregators-client/src/inch_client/mod.rs
@@ -1,1 +1,2 @@
 pub mod client;
+pub mod types;
